requirements.txt
qdrant-client==1.12.0 # This Verson is compatible with the latest Qdrant API
python-dotenv
requests
langchain
langchain-community
langchain-google-genai
langchain-groq
langchain-qdrant
selenium
beautifulsoup4
tqdm
pydantic
rich
unstructured
git+https://github.com/brandonstarxel/chunking_evaluation.git
numpy
scikit-learn
rank_bm25
fastembed-gpu